int -> double

+++ draw.h
@@ -12,7 +12,7 @@ void dl_draw_line_pp(figure *fptr);
 void dl_draw_rect_pp(figure *fptr);
 
 void dl_set_cairo_context(cairo_t *cr);
-void dl_set_preview_coords(int px, int py);
+void dl_set_preview_coords(double px, double py);
 
 int dl_is_need_draw(figure *fptr);
